Added Blueprints nav section
