+++ console/src/routes/infrastructure/templates.tsx
@@ -13,10 +13,6 @@ import { Badge } from "@/components/ui/badge";
 import { TemplateDialog } from "@/components/infrastructure/TemplateDialog";
 import { TemplateInstancesTable } from "@/components/infrastructure/TemplateInstancesTable";
 import { RunLogDialog } from "@/components/infrastructure/RunLogDialog";
-import { StackBlueprintsPanel } from "@/components/infrastructure/StackBlueprintsPanel";
-import { BlueprintDialog, type BlueprintAction } from "@/components/infrastructure/BlueprintDialog";
-import { BlueprintFormDialog, type BlueprintFormAction } from "@/components/infrastructure/BlueprintFormDialog";
-import type { Blueprint, BlueprintGroup } from "@/lib/blueprints";
 import { toast } from "sonner";
 import { api } from "@/lib/api";
 import { useT } from "@/lib/i18n";
@@ -41,14 +37,13 @@ function IconFor({ name }: { name?: string }) {
   return <Ico className="h-5 w-5" />;
 }
 
-type Tab = "jobs" | "blueprints";
+type Tab = "jobs";
 
 function TemplatesPage() {
   const t = useT();
   const qc = useQueryClient();
   const [q, setQ] = useState("");
   const [instSearch, setInstSearch] = useState("");
-  const [bpSearch, setBpSearch] = useState("");
   const [openId, setOpenId] = useState<string | null>(null);
   const [openTemplateInitial, setOpenTemplateInitial] = useState<{
     values?: Record<string, unknown>;
@@ -58,112 +53,7 @@ function TemplatesPage() {
   } | null>(null);
   const [tab, setTab] = useState<Tab>("jobs");
   const [runLogId, setRunLogId] = useState<string | null>(null);
-  const [blueprintPreview, setBlueprintPreview] = useState<{ bp: Blueprint; group: BlueprintGroup } | null>(null);
-  const [blueprintForm, setBlueprintForm] = useState<{ bp: Blueprint; group: BlueprintGroup } | null>(null);
   const [catalogOpen, setCatalogOpen] = useState(false);
-
-  const openBlueprint = (bp: Blueprint, group: BlueprintGroup) => {
-    // Schema-driven blueprints skip the preview and open the generic form directly.
-    if (bp.available && bp.templateId && bp.formSchema && bp.formSchema.length > 0) {
-      setBlueprintForm({ bp, group });
-      return;
-    }
-    setBlueprintPreview({ bp, group });
-  };
-
-  const runBlueprintAction = async (
-    bp: Blueprint,
-    action: BlueprintAction | BlueprintFormAction,
-    valuesOverride?: Record<string, unknown>,
-  ) => {
-    if (!bp.available || !bp.templateId) {
-      toast.error("This blueprint is not yet wired to an executable template.");
-      return;
-    }
-
-    const environment = action === "run-once" ? "sandbox" : "default";
-    const stem = bp.filenameStem || bp.id;
-    const stamp = new Date().toISOString().replace(/[-:T]/g, "").slice(0, 12);
-    const filename = `${stem}-${environment}-${stamp}.yml`;
-    const values = valuesOverride ?? bp.defaults ?? {};
-
-    try {
-      const saved = await api<{ ok: boolean; filename: string; playbook_id?: string; instance_path?: string }>(
-        "POST",
-        `/api/templates/${bp.templateId}/save`,
-        { values, targets: {}, environment, filename },
-      );
-
-      await qc.invalidateQueries({ queryKey: ["template-instances"] });
-      setBlueprintPreview(null);
-      setBlueprintForm(null);
-      setTab("jobs");
-      toast.success(`Job created: ${saved.filename}`);
-
-      if (action === "run-once") {
-        const pbId = saved.playbook_id || saved.filename.replace(/\.ya?ml$/i, "");
-        try {
-          const run = await api<{ executionId?: string; execution_id?: string }>(
-            "POST",
-            `/api/projects/_current/playbooks/${encodeURIComponent(pbId)}/run`,
-            {
-              inventory_files: ["inventory.yml"],
-              ansible_config: "ansible.cfg",
-              play_name: `Blueprint: ${bp.name} [${environment}]`,
-              become: true,
-              strategy: "linear",
-            },
-          );
-          const id = run.executionId || run.execution_id;
-          if (id) {
-            toast.success(`Run queued (${id})`);
-            setRunLogId(String(id));
-          } else {
-            toast.success("Run queued");
-          }
-        } catch (e) {
-          toast.error(`Run failed: ${(e as Error).message}`);
-        }
-      } else {
-        // Create Job: open the saved instance so user can review/tweak variables and re-save.
-        setOpenTemplateInitial({
-          values,
-          filename: saved.filename,
-          environment,
-          instancePath: saved.instance_path,
-        });
-        setOpenId(bp.templateId);
-      }
-    } catch (e) {
-      const err = e as { message?: string; body?: { field_errors?: Record<string, string | string[]> } };
-      const fieldErrors = err?.body?.field_errors;
-      if (fieldErrors && Object.keys(fieldErrors).length > 0) {
-        const details = Object.entries(fieldErrors)
-          .map(([field, msg]) => `${field}: ${Array.isArray(msg) ? msg.join(", ") : msg}`)
-          .join("\n");
-        toast.error("Failed to create job", { description: details });
-      } else {
-        toast.error(`Failed to create job: ${err?.message ?? "unknown error"}`);
-      }
-    }
-  };
-
-
-  const handleBlueprintAction = async (action: BlueprintAction) => {
-    const preview = blueprintPreview;
-    if (!preview) return;
-    await runBlueprintAction(preview.bp, action);
-  };
-
-  const handleBlueprintFormSubmit = async (
-    action: BlueprintFormAction,
-    values: Record<string, unknown>,
-  ) => {
-    const form = blueprintForm;
-    if (!form) return;
-    await runBlueprintAction(form.bp, action, values);
-  };
-
 
   const listQ = useQuery({
     queryKey: ["templates"],
@@ -205,12 +95,6 @@ function TemplatesPage() {
               <Input value={instSearch} onChange={(e) => setInstSearch(e.target.value)} placeholder={t("templates.searchJobs")} className="pl-7 h-8 w-56" />
             </div>
           )}
-          {tab === "blueprints" && (
-            <div className="relative">
-              <Search className="absolute left-2 top-1/2 -translate-y-1/2 h-3.5 w-3.5 text-[var(--color-muted-foreground)]" />
-              <Input value={bpSearch} onChange={(e) => setBpSearch(e.target.value)} placeholder="Search blueprints…" className="pl-7 h-8 w-56" />
-            </div>
-          )}
           <Button variant="outline" size="sm" onClick={() => {
             listQ.refetch();
             qc.invalidateQueries({ queryKey: ["template-instances"] });
@@ -226,7 +110,6 @@ function TemplatesPage() {
       <div className="flex items-center gap-1 border-b border-[var(--color-border)]">
         {([
           ["jobs", t("templates.tab.jobs")],
-          ["blueprints", t("templates.tab.blueprints")],
         ] as const).map(([id, label]) => (
           <button
             key={id}
@@ -243,14 +126,12 @@ function TemplatesPage() {
       </div>
 
       <p className="text-xs text-[var(--color-muted-foreground)]">
-        {tab === "jobs" ? t("templates.tab.jobs.desc") : t("templates.tab.blueprints.desc")}
+        {t("templates.tab.jobs.desc")}
       </p>
 
       {tab === "jobs" && (
         <TemplateInstancesTable filterText={instSearch} onRunLog={(id) => setRunLogId(id)} />
       )}
-
-      {tab === "blueprints" && <StackBlueprintsPanel search={bpSearch} onSelect={openBlueprint} />}
 
       {catalogOpen && (
         <>
@@ -327,24 +208,6 @@ function TemplatesPage() {
           initialFilename={openTemplateInitial?.filename}
           initialEnvironment={openTemplateInitial?.environment}
           instancePath={openTemplateInitial?.instancePath}
-        />
-      )}
-
-      {blueprintPreview && (
-        <BlueprintDialog
-          blueprint={blueprintPreview.bp}
-          group={blueprintPreview.group}
-          onClose={() => setBlueprintPreview(null)}
-          onAction={handleBlueprintAction}
-        />
-      )}
-
-      {blueprintForm && (
-        <BlueprintFormDialog
-          blueprint={blueprintForm.bp}
-          group={blueprintForm.group}
-          onClose={() => setBlueprintForm(null)}
-          onSubmit={handleBlueprintFormSubmit}
         />
       )}
